backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic
sqlalchemy==2.0.23
python-multipart==0.0.6
pytest==7.4.3
beautifulsoup4==4.12.3
bs4==0.0.2
requests==2.32.3
numpy==2.2.1
urllib3==2.3.0
openai==1.59.7
firecrawl-py==1.9.0
httpx
numpy